@@ -10,30 +10,31 @@
     "logs": "firebase functions:log"
   },
   "engines": {
-    "node": "12"
+    "node": "14"
   },
+  "type": "module",
   "main": "lib/index.js",
   "dependencies": {
     "@types/jsdom": "^16.2.13",
-    "@types/node-fetch": "^2.5.12",
+    "@types/node-fetch": "^3.0.2",
     "amazon-paapi": "^1.0.5",
-    "firebase-admin": "^9.11.0",
-    "firebase-functions": "^3.14.1",
-    "jsdom": "^16.7.0",
-    "node-fetch": "^2.6.1",
+    "firebase-admin": "^9.12.0",
+    "firebase-functions": "^3.15.7",
+    "jsdom": "^18.0.0",
+    "node-fetch": "^3.0.0",
     "paapi5-nodejs-sdk": "^1.1.0"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^4.29.0",
-    "@typescript-eslint/parser": "^4.29.0",
+    "@typescript-eslint/eslint-plugin": "^4.33.0",
+    "@typescript-eslint/parser": "^4.33.0",
     "eslint": "^7.32.0",
     "eslint-config-google": "^0.14.0",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.24.0",
-    "firebase-functions-test": "^0.3.1",
-    "prettier": "^2.3.2",
-    "tsc-watch": "^4.4.0",
-    "typescript": "^4.3.5"
+    "eslint-plugin-import": "^2.24.2",
+    "firebase-functions-test": "^0.3.3",
+    "prettier": "^2.4.1",
+    "tsc-watch": "^4.5.0",
+    "typescript": "^4.4.3"
   },
   "private": true
 }
